fix: little adjustments
